requirements.txt
# Core dependencies
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0

# CLI and utilities
click>=8.1.0
requests>=2.31.0
lxml>=4.9.0

# Development dependencies
black>=23.0.0
pytest>=7.4.0
